Refined Routes and Added md files
